# [Enhancement] Support opening preaggregation when the case when/if contains a constant zero