## What type of PR is this：
- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

## Which issues of this PR fixes ：
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #SR-15810

## Problem Summary(Required) ：
<!-- (Please describe the changes you have made. In which scenarios will this bug be triggered and what measures have you taken to fix the bug?) -->
```
CREATE TABLE IF NOT EXISTS test_agg_2 (
  k1 int(11) NULL,
  k2 int(11) NULL,
  k3 int(11) NULL,
  v1 int SUM NULL,
  v2 bigint SUM NULL
)

select sum(case when k1 = 1 then v1 else 0 end),
       sum(case when k1 = 1 then v2 else 0 end),
       sum(case when k1 = 1 then 0 else v1 end),
       sum(case when k1 = 1 then 0 else v2 end)
from test_agg_2;
```
The case when contains a constant zero. The preaggregation can be opened. Adding an isConstantZero() function to help to address the problem.

I test the performance using catalog_sales.
It contains 19 aggregate key columns, the number of rows is 144992165.
There are 146 rowsets, each has 1 million rows.
```
mysql> desc catalog_sales;
+--------------------------+---------------+------+-------+---------+-------+
| Field                    | Type          | Null | Key   | Default | Extra |
+--------------------------+---------------+------+-------+---------+-------+
| cs_item_sk               | bigint        | YES  | true  | NULL    |       |
| cs_order_number          | bigint        | YES  | true  | NULL    |       |
| cs_sold_date_sk          | bigint        | YES  | true  | NULL    |       |
| cs_sold_time_sk          | bigint        | YES  | true  | NULL    |       |
| cs_ship_date_sk          | bigint        | YES  | true  | NULL    |       |
| cs_bill_customer_sk      | bigint        | YES  | true  | NULL    |       |
| cs_bill_cdemo_sk         | bigint        | YES  | true  | NULL    |       |
| cs_bill_hdemo_sk         | bigint        | YES  | true  | NULL    |       |
| cs_bill_addr_sk          | bigint        | YES  | true  | NULL    |       |
| cs_ship_customer_sk      | bigint        | YES  | true  | NULL    |       |
| cs_ship_cdemo_sk         | bigint        | YES  | true  | NULL    |       |
| cs_ship_hdemo_sk         | bigint        | YES  | true  | NULL    |       |
| cs_ship_addr_sk          | bigint        | YES  | true  | NULL    |       |
| cs_call_center_sk        | bigint        | YES  | true  | NULL    |       |
| cs_catalog_page_sk       | bigint        | YES  | true  | NULL    |       |
| cs_ship_mode_sk          | bigint        | YES  | true  | NULL    |       |
| cs_warehouse_sk          | bigint        | YES  | true  | NULL    |       |
| cs_promo_sk              | bigint        | YES  | true  | NULL    |       |
| cs_quantity              | bigint        | YES  | true  | NULL    |       |
| cs_wholesale_cost        | decimal(38,2) | YES  | false | NULL    |       |
| cs_list_price            | decimal(38,2) | YES  | false | NULL    |       |
| cs_sales_price           | decimal(38,2) | YES  | false | NULL    |       |
| cs_ext_discount_amt      | decimal(38,2) | YES  | false | NULL    |       |
| cs_ext_sales_price       | decimal(38,2) | YES  | false | NULL    |       |
| cs_ext_wholesale_cost    | decimal(38,2) | YES  | false | NULL    |       |
| cs_ext_list_price        | decimal(38,2) | YES  | false | NULL    |       |
| cs_ext_tax               | decimal(38,2) | YES  | false | NULL    |       |
| cs_coupon_amt            | decimal(38,2) | YES  | false | NULL    |       |
| cs_ext_ship_cost         | decimal(38,2) | YES  | false | NULL    |       |
| cs_net_paid              | decimal(38,2) | YES  | false | NULL    |       |
| cs_net_paid_inc_tax      | decimal(38,2) | YES  | false | NULL    |       |
| cs_net_paid_inc_ship     | decimal(38,2) | YES  | false | NULL    |       |
| cs_net_paid_inc_ship_tax | decimal(38,2) | YES  | false | NULL    |       |
| cs_net_profit            | decimal(38,2) | YES  | false | NULL    |       |
+--------------------------+---------------+------+-------+---------+-------+
```

The test query is
```
select sum(case when cs_item_sk > 0 then cs_net_profit else 0.0 end) from catalog_sales;
```
The cold query latency reduce from 80s -> 3.6s
The warm query latency reduce from 10s -> 0.2s
The peak cpu consumption reduce from 9000% -> 700% 

## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto backported to target branch
  - [x] 3.0
  - [x] 2.5
  - [ ] 2.4
  - [ ] 2.3