@wanpengfei-git 2023-03-14
#### [FE PR Coverage Check]

:heart_eyes: **pass** : 28 / 31 (90.32%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/sql/optimizer/operator/scalar/ConstantOperator.java|19|22|86.36%|[197, 198, 206]|
|:large_blue_circle:|com/starrocks/sql/optimizer/operator/scalar/ScalarOperator.java|1|1|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/tree/PreAggregateTurnOnRule.java|8|8|100.00%|[]|