@wanpengfei-git 2023-03-15
run starrocks_admit_test